requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.3
lxml>=4.9.3
schedule>=1.2.0
smtplib2>=0.2.1
fake-useragent>=1.4.0
python-dotenv>=1.0.0
colorama>=0.4.6